add gitignore and switch back to int

+++ include/evaluate.h
@@ -39,8 +39,8 @@
 /* https://www.chessprogramming.org/Evaluation                               */
 /* https://www.chessprogramming.org/Piece-Square_Tables                      */
 /* https://www.chessprogramming.org/Mobility                                 */
-int evaluate(const struct position *pos, short **table);
+int evaluate(const struct position *pos, int **table);
 
-int	init_pst(short ***table);
+int	init_pst(int ***table);
 
 #endif
